src/backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
passlib[bcrypt]==1.7.4

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Data Processing
# pandas==2.2.0  # Temporarily disabled due to Python 3.13 compatibility
# numpy==1.26.0   # Temporarily disabled due to Python 3.13 compatibility
python-dateutil==2.8.2

# HTTP Client & CORS
httpx==0.25.2
# python-cors==1.7.0  # Package name incorrect, will fix later

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# File Processing
openpyxl==3.1.2  # For Excel file support if needed

# Utilities
python-slugify==8.0.1
click==8.1.7
